refine gatsby-node and draw out functions
